refactor: remove useServerSelection hook and consolidate server management into usePlexManagement
- Deleted the useServerSelection hook and its associated state management logic.
- Updated index.ts to export usePlexManagement instead of useServerSelection.
- Enhanced usePlexManagement to handle server selection, authentication, and connection testing.
- Introduced new components: PlexAuthSection, PlexConnectionCard, and PlexServerSection for improved UI.
- Updated queries to include user-specific cache keys for better data handling.
- Improved error handling and messaging throughout the application.
- Removed deprecated serverCache utility.
