+++ frontend/src/plex/utilities/errors.ts
@@ -1,86 +1,66 @@
 import { PLEX_ERROR_CODES, type PlexErrorCode } from "@/plex/constants/auth";
 
-interface AxiosErrorResponse {
-  data?: {
-    error?: string;
-    code?: string;
-  };
-  status?: number;
-}
-
-interface AxiosError extends Error {
-  response?: AxiosErrorResponse;
-  isAxiosError?: boolean;
-}
-
 export interface PlexError {
   message: string;
   code: PlexErrorCode;
   isRetryable?: boolean;
 }
 
-export function createPlexError(
-  message: string,
-  code: PlexErrorCode,
-  isRetryable = false,
-): PlexError {
-  return { message, code, isRetryable };
+interface AxiosError extends Error {
+  response?: {
+    data?: { error?: string; code?: string };
+    status?: number;
+  };
+  isAxiosError?: boolean;
 }
 
-function isAxiosError(error: unknown): error is AxiosError {
-  return typeof error === "object" && error !== null && "isAxiosError" in error;
-}
+export function parseAxiosError(error: unknown): PlexError {
+  // Handle axios errors
+  if (error && typeof error === "object" && "isAxiosError" in error) {
+    const axiosError = error as AxiosError;
+    const responseData = axiosError.response?.data;
+    const errorCode =
+      (responseData?.code as PlexErrorCode) ||
+      PLEX_ERROR_CODES.CONNECTION_ERROR;
+    const message =
+      responseData?.error || axiosError.message || "Connection error";
+    const isRetryable = axiosError.response?.status
+      ? axiosError.response.status >= 500
+      : false;
 
-function isError(error: unknown): error is Error {
-  return error instanceof Error;
-}
-
-export function parseError(error: unknown): PlexError {
-  // Handle axios-like errors (for backward compatibility)
-  if (isAxiosError(error)) {
-    const responseData = error.response?.data;
-    const errorCode = responseData?.code as PlexErrorCode;
-    const errorMessage = responseData?.error || error.message;
-
-    return createPlexError(
-      errorMessage || "An unknown error occurred",
-      errorCode || PLEX_ERROR_CODES.CONNECTION_ERROR,
-      error.response?.status ? error.response.status >= 500 : false,
-    );
+    return { message, code: errorCode, isRetryable };
   }
 
   // Handle generic errors
-  if (isError(error)) {
-    return createPlexError(error.message, PLEX_ERROR_CODES.CONNECTION_ERROR);
+  if (error instanceof Error) {
+    return {
+      message: error.message,
+      code: PLEX_ERROR_CODES.CONNECTION_ERROR,
+      isRetryable: false,
+    };
   }
 
-  // Handle unknown errors
-  return createPlexError(
-    "An unknown error occurred",
-    PLEX_ERROR_CODES.CONNECTION_ERROR,
-  );
+  // Fallback
+  return {
+    message: "An unknown error occurred",
+    code: PLEX_ERROR_CODES.CONNECTION_ERROR,
+    isRetryable: false,
+  };
 }
 
-// Keep the old function for backward compatibility
-export const parseAxiosError = parseError;
-
 export function getErrorMessage(error: PlexError): string {
-  const codeMessages: Record<PlexErrorCode, string> = {
+  const messages: Record<PlexErrorCode, string> = {
     [PLEX_ERROR_CODES.PIN_EXPIRED]:
-      "Authentication PIN has expired. Please try again.",
+      "Authentication PIN expired. Please try again.",
     [PLEX_ERROR_CODES.AUTH_TIMEOUT]:
       "Authentication timed out. Please try again.",
-    [PLEX_ERROR_CODES.CONNECTION_ERROR]:
-      "Unable to connect to Plex. Please check your connection.",
-    [PLEX_ERROR_CODES.INVALID_TOKEN]:
-      "Invalid authentication token. Please re-authenticate.",
-    [PLEX_ERROR_CODES.SERVER_NOT_FOUND]:
-      "Plex server not found or unavailable.",
-    [PLEX_ERROR_CODES.UNAUTHORIZED]:
-      "Unauthorized access. Please re-authenticate.",
+    [PLEX_ERROR_CODES.CONNECTION_ERROR]: "Unable to connect to Plex.",
+    [PLEX_ERROR_CODES.INVALID_TOKEN]: "Invalid token. Please re-authenticate.",
+    [PLEX_ERROR_CODES.SERVER_NOT_FOUND]: "Plex server not found.",
+    [PLEX_ERROR_CODES.UNAUTHORIZED]: "Unauthorized. Please re-authenticate.",
   };
 
   return (
-    error.message || codeMessages[error.code] || "An unexpected error occurred"
+    error.message || messages[error.code] || "An unexpected error occurred"
   );
 }
